% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_src_dir.R
\name{build_src_dir}
\alias{build_src_dir}
\title{FUNCTION_TITLE}
\usage{
build_src_dir(pkg, version = NA, date = NA)
}
\arguments{
\item{pkg}{PARAM_DESCRIPTION}

\item{version}{PARAM_DESCRIPTION, Default: NA}

\item{date}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
