% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_children.R
\name{find_children}
\alias{find_children}
\title{FUNCTION_TITLE}
\usage{
find_children(package, uri)
}
\arguments{
\item{package}{PARAM_DESCRIPTION}

\item{uri}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
