% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archive.R
\name{get_archive}
\alias{get_archive}
\title{FUNCTION_TITLE}
\usage{
get_archive(package, min_version = NA, min_date = NA)
}
\arguments{
\item{package}{PARAM_DESCRIPTION}

\item{min_version}{PARAM_DESCRIPTION, Default: NA}

\item{min_date}{PARAM_DESCRIPTION, Default: NA}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
